requirements.txt
aiohttp==3.6.2
appdirs==1.4.4
async-timeout==3.0.1
attrs==19.3.0
beautifulsoup4==4.9.0
certifi==2020.4.5.1
chardet==3.0.4
discord.py==1.3.3
distlib==0.3.0
filelock==3.0.12
idna==2.9
importlib-metadata==1.6.0
multidict==4.7.5
pipenv==2018.11.26
six==1.14.0
soupsieve==2.0
virtualenv==20.0.20
virtualenv-clone==0.5.4
websockets==8.1
yarl==1.4.2
zipp==3.1.0
